feat: updated styling for header and buttons
