+++ services/web-app/src/styles/updateApiKey.css
@@ -49,28 +49,10 @@ input {
 }
 
 .Button {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 4px;
-  padding: 0 15px;
-  font-size: 15px;
-  line-height: 1;
-  font-weight: 500;
-  height: 35px;
-  width: max-content;
+  border-radius: 6px;
+  padding: 10px;
 }
-.Button.violet {
-  background-color: black;
-  color: white;
-}
-.Button.violet:hover {
-  text-decoration: underline;
-  background-color: black;
-}
-.Button.violet:focus {
-  box-shadow: 0 0 0 2px black;
-}
+
 .Button.green {
   background-color: var(--green-4);
   color: var(--green-11);
